requirements.txt
PyQt5>=5.15.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
